update id Long to Integer

+++ src/main/java/com/bootcamp/itexperts/repositories/AccountRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.bootcamp.itexperts.models.AccountModel;
 
 @Repository
-public interface AccountRepository extends JpaRepository<AccountModel, Long>{
+public interface AccountRepository extends JpaRepository<AccountModel, Integer>{
 	
 }
